requirements.txt
Django==2.2.13
python-dotenv==0.13.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
line-bot-sdk==1.19.0